Allow TypeScript file extension to pass the file exists nodejs runtime check
Out of the box the Base runtime class will check to see if a given source handler exists. This uses the "extension" value for the given runtime (e.g. `this.extension = ".js"` in the case of nodejs. Because this value is hardcoded to a single value, if transpiling TypeScript using the `serverless-plugin-typescript` plugin for example the `isValidFile` check will fail because there will be no handler file with the given path because it is just assumed that the file will end with ".js".

This commit adds a check specifically for a `.ts` file if the check for the `.js` file fails.